src/com/ch01/testprogram/temp/TrashAndTakeOut.java
package com.ch01.testprogram.temp;

/**
 * TrashAndTakeOut类用于演示垃圾和取出操作。
 */
public class TrashAndTakeOut {

    /**
     * 计算垃圾和取出操作的结果。
     *
     * @param x 输入参数
     * @return 操作结果
     */
    public static int trash(int x) {
        int m;
        int n;
        m = 0;
        if (x > 0) {
            m = 4;
        }
        if (x > 5) {
            n = 3 * m;
        } else {
            n = 4 * m;
        }
        // 调用takeOut方法并返回结果
        int o = takeOut(m, n);
        return o;
    }

    /**
     * 执行取出操作。
     *
     * @param a 输入参数1
     * @param b 输入参数2
     * @return 取出操作的结果
     */
    public static int takeOut(int a, int b) {
        int d;
        int e;
        // 计算d的值
        d = 42 * a;
        if (a > 0) {
            // 如果a大于0，则计算e的值
            e = 2 * b + d;
        } else {
            // 如果a不大于0，则计算e的值
            e = b + d;
        }
        return e;
    }
}